Further improvements to conversion to optimized lookup format
This speeds up conversion a bit more, and also improves packing
in one case by around 25% (implying that this commit *does* change the
output of the conversion, but in a functionally equivalent way).

The space gain was mainly accomplished by changing what happens when we fail
to find a suitable index table location several times a row starting in the
same place. We used to jump up to the previous successful location and its
indices, now we keep trying to fill in the gaps between those first.

This commit also removes the state-relabeling facility, because it didn't seem
to be doing anything (perhaps something in HfstBasicTransducer has changed;
it seems state numbers can no longer be non-contiguous). So we now assume
states are contiguous and iterated in order. If this assumption is broken,
this needs to be changes. But as I said earlier, looks like relabeling wasn't
doing anything now anyway.


git-svn-id: svn+ssh://svn.code.sf.net/p/hfst/code/trunk/hfst3@4136 941e2c2b-deac-454f-805a-451daa25f33c