Modify minimization of weighted transducers so that negative weights are handled correctly.
After epsilon removal, all states and transitions of the transducer are iterated over to find the lowest weight.
If the lowest weight is negative, its opposite is added to all weights in the transducer.
Then the transducer is determinized and minimized normally.
Finally, if the original lowest weight was negative, it is added to all weights in the transducer.

This procedure effectively removes negative weights from the transducer before determinization and adds them again in the end.
It is needed because OpenFst's determinization algorithm cannot handle negative weights correctly.

Running 'make check' in directory fsmbook-examples takes approximately 3 minutes before and after the changes, so it seems that the slowdown should not be an issue.

git-svn-id: svn+ssh://svn.code.sf.net/p/hfst/code/trunk/hfst3@4439 941e2c2b-deac-454f-805a-451daa25f33c