Make binary command line tools convert mismatching input streams into common format.
If the input streams have a different implementation type, convert transducers from the second stream into the type of the first stream.
If information can be lost, i.e. weighted format is to be converted into unweighted format, convert the streams vice versa.
Option --do-not-convert, in short form -C, can be used to disallow conversion and return with a failure value.

git-svn-id: svn+ssh://svn.code.sf.net/p/hfst/code/trunk/hfst3@4570 941e2c2b-deac-454f-805a-451daa25f33c