Revert changes to hfst-proc made in revision 4427. This will (temporarily) fix bug #328.
Change typedef ProcResults back to set<string> from vector<string>.
This will remove the duplicates, but the results will again be listed in another order than they were inserted.
vector<string> is probably be a better structure for storing the results, but it requires more changes in the code than just changing the typedef.

git-svn-id: svn+ssh://svn.code.sf.net/p/hfst/code/trunk/hfst3@4579 941e2c2b-deac-454f-805a-451daa25f33c